requirements.txt
python-dotenv>=0.19.2
spotipy>=2.19.0
